+++ src/utils.js
@@ -72,7 +72,7 @@ const utils = {
     const isAlphanumeric =
       ev.key.length === 1 ||
       ["Enter", "Space", "Tab", "Escape", "Backspace", "Delete"].includes(
-        ev.key
+        ev.key,
       ) ||
       ev.key.startsWith("F") || // Function keys
       ["Home", "End", "PageUp", "PageDown", "Insert"].includes(ev.key);
@@ -137,7 +137,7 @@ const utils = {
           reject(
             typeof ret.error === "string"
               ? new ErrorWithMoreInfo(ret.error, ret)
-              : ret.error
+              : ret.error,
           );
         } else {
           resolve(ret);
@@ -173,12 +173,12 @@ const utils = {
       chrome.tabs.sendMessage(tabId, data, (ret) => {
         if (ret?.error) {
           reject(
-            typeof ret.error === "string" ? new Error(ret.error) : ret.error
+            typeof ret.error === "string" ? new Error(ret.error) : ret.error,
           );
         } else {
           resolve(ret);
         }
-      })
+      }),
     );
     if (callback) {
       return p.then(callback);
@@ -196,8 +196,8 @@ const utils = {
           window.dictionariezBackgroundListeners[request.type]?.(
             request,
             sender,
-            sendResponse
-          )
+            sendResponse,
+          ),
         );
       } else {
         window.dictionariezBackgroundListeners[type] = callback;
@@ -292,7 +292,7 @@ const utils = {
         str
           .split(/\s/)
           .filter(
-            (w) => w.length > 1 && !simpleStopWords.includes(w.toLowerCase())
+            (w) => w.length > 1 && !simpleStopWords.includes(w.toLowerCase()),
           ).length >= minLength
       );
     }
@@ -384,7 +384,7 @@ const utils = {
         method: "GET",
         credentials,
       }),
-      5000
+      5000,
     ).then((resp) => {
       if (!resp.ok) {
         const err = new Error(resp.statusText);
@@ -402,7 +402,7 @@ const utils = {
         method: "GET",
         credentials,
       }),
-      5000
+      5000,
     ).then((resp) => {
       if (!resp.ok) {
         const err = new Error(resp.statusText);
@@ -412,6 +412,27 @@ const utils = {
 
       return resp.json();
     });
+  },
+  checkEditable(element) {
+    let curNode = element;
+    while (curNode) {
+      if (
+        curNode.isContentEditable ||
+        ["input", "textarea"].includes(curNode.nodeName.toLowerCase())
+      ) {
+        return true;
+      }
+      curNode = curNode.parentElement;
+    }
+    // check the direct children of the node, sometimes the editor could be wrapped by a div.
+    for (const node of element?.children || []) {
+      if (
+        node.isContentEditable ||
+        ["input", "textarea"].includes(node.nodeName.toLowerCase())
+      ) {
+        return true;
+      }
+    }
   },
 };